group and role

+++ src/main/java/com/keysoft/ecommerce/service/impl/GroupServiceImpl.java
@@ -1,17 +1,22 @@
 package com.keysoft.ecommerce.service.impl;
 
 import com.keysoft.ecommerce.dto.GroupDTO;
+import com.keysoft.ecommerce.dto.RoleDTO;
 import com.keysoft.ecommerce.model.Group;
 import com.keysoft.ecommerce.repository.GroupRepository;
 import com.keysoft.ecommerce.service.GroupService;
+import com.keysoft.ecommerce.util.CodeHelper;
 import lombok.extern.slf4j.Slf4j;
+import org.hibernate.Hibernate;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.util.StringUtils;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Set;
 
 @Service
 @Slf4j
@@ -32,31 +37,68 @@ public class GroupServiceImpl implements GroupService {
 
     @Override
     @Transactional(rollbackFor = {Exception.class, Throwable.class})
-    public boolean save(GroupDTO GroupDTO) {
-        Group Group;
-        if(GroupDTO.getId() != null) {
-            Group = GroupRepository.findById(GroupDTO.getId()).orElse(null);
-            Group.setName(GroupDTO.getName());
-        }else {
-            Group = modelMapper.map(GroupDTO, Group.class);
+    public boolean save(GroupDTO groupDTO) {
+        log.info("SERVICE : SAVE GROUP, GROUP: {}", groupDTO);
+        boolean isCheck = checkGroupExist(groupDTO.getName());
+        if(isCheck){
+            throw  new IllegalStateException("Nhóm người dùng đã tồn tại");
         }
-        return GroupRepository.save(Group).getId() != null;
+
+        Set<RoleDTO> roles;
+
+        if(groupDTO.getId() != null && GroupRepository.findById(groupDTO.getId()).isEmpty())
+            throw new NullPointerException("Nhóm người dùng không tồn tại");
+
+        if (StringUtils.hasText(groupDTO.getName()) && groupDTO.getRoles() != null) {
+            roles = Set.copyOf(groupDTO.getRoles());
+        } else {
+            throw new NullPointerException("Không có quyền của nhóm hợp lệ");
+        }
+
+        groupDTO.setRoles(roles);
+        groupDTO.setCode(CodeHelper.spawnCodeFromName(groupDTO.getName()));
+
+        return GroupRepository.save(modelMapper.map(groupDTO, Group.class)).getId() != null;
+    }
+
+    public Boolean checkGroupExist(String name) {
+        log.info("SERVICE PROCESS: CHECK GROUP NAME USED, NAME: {}", name);
+
+        Group group = GroupRepository.findByName(name).orElse(new Group());
+        return group.getId() != null;
     }
 
     @Override
     public GroupDTO get(Long id) {
-        return modelMapper.map(GroupRepository.findById(id), GroupDTO.class);
+        GroupDTO groupDTO = modelMapper.map(GroupRepository.findById(id).orElse(new Group()), GroupDTO.class);
+
+        List<RoleDTO> roles = List.copyOf(groupDTO.getRoles());
+        List<Long> roleIds = new ArrayList<>();
+
+        roles.forEach(item -> roleIds.add(item.getId()));
+        groupDTO.setListId(roleIds);
+
+        return groupDTO;
     }
 
     @Override
     @Transactional(rollbackFor = {Exception.class, Throwable.class})
     public boolean delete(Long id) {
-        try {
+
+        Group selected = GroupRepository.findById(id).orElse(new Group());
+
+        if (selected.getId() == null) {
+            throw new IllegalStateException("Không tìm được nhóm người dùng");
+        }
+
+        Hibernate.initialize(selected.getUsers());
+
+        if (selected.getUsers().isEmpty()) {
             GroupRepository.deleteById(id);
-            return true;
-        } catch (Exception e) {
-            e.printStackTrace();
+        } else {
             return false;
         }
+
+        return true;
     }
 }
